+++ examples/simulation/plot_simulate_evoked_data.py
@@ -25,7 +25,7 @@ print(__doc__)
 data_path = sample.data_path()
 
 raw = mne.io.read_raw_fif(data_path + '/MEG/sample/sample_audvis_raw.fif')
-proj = mne.read_proj(data_path + '/MEG/sample/sample_audvis_ecg_proj.fif')
+proj = mne.read_proj(data_path + '/MEG/sample/sample_audvis_ecg-proj.fif')
 raw.info['projs'] += proj
 raw.info['bads'] = ['MEG 2443', 'EEG 053']  # mark bad channels