+++ examples/time_frequency/plot_temporal_whitening.py
@@ -24,7 +24,7 @@ print(__doc__)
 data_path = sample.data_path()
 
 raw_fname = data_path + '/MEG/sample/sample_audvis_raw.fif'
-proj_fname = data_path + '/MEG/sample/sample_audvis_ecg_proj.fif'
+proj_fname = data_path + '/MEG/sample/sample_audvis_ecg-proj.fif'
 
 raw = mne.io.read_raw_fif(raw_fname)
 proj = mne.read_proj(proj_fname)