+++ examples/visualization/plot_ssp_projs_topomaps.py
@@ -19,10 +19,12 @@ print(__doc__)
 
 data_path = sample.data_path()
 
-ecg_fname = data_path + '/MEG/sample/sample_audvis_ecg_proj.fif'
+ecg_fname = data_path + '/MEG/sample/sample_audvis_ecg-proj.fif'
 ave_fname = data_path + '/MEG/sample/sample_audvis-ave.fif'
 
-evoked = read_evokeds(ave_fname, condition='Left Auditory')
+evoked = read_evokeds(ave_fname, condition='Left Auditory',
+                      baseline=(None, 0.))
+
 projs = read_proj(ecg_fname)
 evoked.add_proj(projs)